+++ source/extensions/isaacsim.physics.newton/docs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.1.1] - 2026-02-06
+
+### Added
+- Capability check
+
+### Removed
+- Data handling from the physics umbrella
+
 ## [0.1.0] - 2025-12-23
 
 ### Added